app: Add new Electron based app
The old one was removed with Lektor 3.2 because it was using an old
Electron build chain and didn't see many updates in it's lifecycle. This
is using the widespread and officially recommended electron-builder and
electron-webpack to build a bundle for not just Mac but also Windows. My
intention is to make this app also viable on Linux.